refactor: 보안을 위해 pendingCredentials 제거 후 로그인 상태 관리 #17
