+++ src/store/auth.ts
@@ -1,18 +1,14 @@
 import { create } from 'zustand'
 import { persist } from 'zustand/middleware'
 import { UserInfo } from '../types/models/user'
-import { LoginUserRequest } from '../types/api/user';
 
 interface AuthState {
   user: UserInfo | null;
   isAuthenticated: boolean;
   accessToken: string | null;
-  pendingCredentials: LoginUserRequest | null;
   
   setUser: (userData: UserInfo, token: string) => void;
   setAccessToken: (token: string) => void;
-  setPendingCredentials: (credentials: LoginUserRequest) => void;
-  clearPendingCredentials: () => void;
   logout: () => void;
 }
 
@@ -22,7 +18,6 @@ export const useAuthStore = create<AuthState>()(
       user: null,
       isAuthenticated: false,
       accessToken: null,
-      pendingCredentials: null,
 
       setUser: (userData, token) => set({ 
         user: userData,
@@ -34,19 +29,10 @@ export const useAuthStore = create<AuthState>()(
         accessToken: token
       }),
 
-      setPendingCredentials: (credentials) => set({
-        pendingCredentials: credentials
-      }),
-
-      clearPendingCredentials: () => set({
-        pendingCredentials: null
-      }),
-
       logout: () => set({ 
         user: null, 
         isAuthenticated: false,
         accessToken: null,
-        pendingCredentials: null
       }),
     }),
     {
